% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDecision.R
\name{addDecision}
\alias{addDecision}
\title{Creates a "decision" object in XML.}
\usage{
addDecision(decision, entity = c(), file = "")
}
\arguments{
\item{decision}{A description of the decision, given as a string.}

\item{entity}{A list of entities upon which the decision was based.}

\item{file}{The filename to write the decision object to. If blank, the XML will be written to the console.}
}
\value{
A structured XML object containing the decision
}
\description{
\code{addDecision} creates an "decision" object in XML.
}
\examples{

addDecision(decision="Final model",
  entity=list("/models/Step7/GOF_MLX_1.png",
     "/models/Step7/GOF_MLX_2.png",
     "/models/Step7/GOF_MLX_3.png",
     "/models/Step7/Step7.SO.xml"),
  file="dec1.xml")

}

