% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivities.R
\name{getActivities}
\alias{getActivities}
\title{Create a table of activities from PROV-JSON input.}
\usage{
getActivities(z)
}
\arguments{
\item{z}{A PROV-JSON data frame.}
}
\value{
A \code{data.frame} containing \code{activity}, \code{startTime}, \code{description}, 
and \code{displayName} fields.
}
\description{
\code{getActivities} returns a data frame containing a list of activities and pertinent
provenance information, extracted from a DDMoRe PROV-JSON document.
}
\examples{
### Reads in a sequence of JSON files from dir
loadOutput <- function(files) {
jList <- list()
for(f in files) {
  jList[[basename(f)]] <- fromJSON(txt = f)
}
jList
}

dir   <- "/data/jsonDocs"
jlist <- loadOutput(file.path(dir, dir(dir)))

j1 <- lapply(jlist, function(x) {
  z <- getActivities(x)
  z
})
jAct       <- do.call(rbind, j1)
jAct$index <- row.names(jAct)
}
\seealso{
\url{https://www.w3.org/TR/prov-o/#Activity}
}

