% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAssociatedWith.R
\name{getAssociatedWith}
\alias{getAssociatedWith}
\title{Create a table of "wasAssociatedWith" relationships from PROV-JSON input.}
\usage{
getAssociatedWith(z)
}
\arguments{
\item{z}{A PROV-JSON data frame.}
}
\value{
A \code{data.frame} containing \code{associatedWith}, \code{activity}, and \code{plan} fields.
}
\description{
\code{getAssociatedWith} returns a data frame containing the PROV-O "associatedWith" relationships extracted from a DDMoRe PROV-JSON document. An activity association is an assignment of responsibility to an agent for an activity, indicating that the agent had a role in the activity. It further allows for a plan to be specified, which is the plan intended by the agent to achieve some goals in the context of this activity.
}
\examples{
### Reads in a sequence of JSON files from dir
loadOutput <- function(files) {
jList <- list()
for(f in files) {
  jList[[basename(f)]] <- fromJSON(txt = f)
}
jList
}

dir   <- "/data/jsonDocs"
jlist <- loadOutput(file.path(dir, dir(dir)))

j1 <- lapply(jlist, function(x) {
  z <- getAssociatedWith(x)
  z
})
jAss       <- do.call(rbind, j1)
jAss$index <- row.names(jAss)

}
\seealso{
\url{https://www.w3.org/TR/prov-o/#wasAssociatedWith}
}

