% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDerivedFrom.R
\name{getDerivedFrom}
\alias{getDerivedFrom}
\title{Create a table of "wasDerivedFrom" relationships from PROV-JSON input.}
\usage{
getDerivedFrom(z)
}
\arguments{
\item{z}{A PROV-JSON data frame.}
}
\value{
A \code{data.frame} containing \code{derivedFrom}, \code{generatedEntity}, and \code{usedEntity} fields.
}
\description{
\code{getDerivedFrom} returns a data frame containing the PROV-O "wasDerivedFrom" relationships extracted from a DDMoRe PROV-JSON document. A derivation is a transformation of an entity into another, an update of an entity resulting in a new one, or the construction of a new entity based on a pre-existing entity.
}
\examples{
### Reads in a sequence of JSON files from dir
loadOutput <- function(files) {
jList <- list()
for(f in files) {
  jList[[basename(f)]] <- fromJSON(txt = f)
}
jList
}

dir   <- getwd()
jlist <- loadOutput(file.path(dir, dir(dir)))

j1 <- lapply(jlist, function(x) {
  z <- getAssociatedWith(x)
  z
})
jDeriv       <- do.call(rbind, j1)
jDeriv$index <- row.names(jDeriv)

}
\seealso{
\url{https://www.w3.org/TR/prov-o/#wasDerivedFrom}
}

