% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEntities.R
\name{getEntities}
\alias{getEntities}
\title{Create a table of entities from PROV-JSON input.}
\usage{
getEntities(z)
}
\arguments{
\item{z}{A PROV-JSON data frame.}
}
\value{
A \code{data.frame} containing \code{entity}, \code{location}, \code{time},
\code{owningBundle}, and \code{type} fields.
}
\description{
\code{getEntities} returns a data frame containing a list of entities and pertinent
provenance information, extracted from a DDMoRe PROV-JSON document.
}
\examples{
### Reads in a sequence of JSON files from dir
loadOutput <- function(files) {
jList <- list()
for(f in files) {
  jList[[basename(f)]] <- fromJSON(txt = f)
}
jList
}

dir   <- "/data/jsonDocs"
jlist <- loadOutput(file.path(dir, dir(dir)))

j1 <- lapply(jlist, function(x) {
  z <- getEntities(x)
  z
})
jEnt       <- do.call(rbind, j1)
jEnt$index <- row.names(jEnt)
}
\seealso{
\url{https://www.w3.org/TR/prov-o/#Entity}
}

