% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGeneratedBy.R
\name{getGeneratedBy}
\alias{getGeneratedBy}
\title{Create a table of "wasGeneratedBy" relationships from PROV-JSON input.}
\usage{
getGeneratedBy(z)
}
\arguments{
\item{z}{A PROV-JSON data frame.}
}
\value{
A \code{data.frame} containing \code{gen}, \code{activity}, and \code{entity} fields.
}
\description{
\code{getGeneratedBy} returns a data frame containing the PROV-O "wasGeneratedBy" relationships extracted from a DDMoRe PROV-JSON document. Generation is the completion of production of a new entity by an activity. This entity did not exist before generation and becomes available for usage after this generation.
}
\examples{
### Reads in a sequence of JSON files from dir
loadOutput <- function(files) {
jList <- list()
for(f in files) {
  jList[[basename(f)]] <- fromJSON(txt = f)
}
jList
}

dir   <- "/data/jsonDocs"
jlist <- loadOutput(file.path(dir, dir(dir)))

j1 <- lapply(jlist, function(x) {
  z <- getGeneratedBy(x)
  z
})
jGen       <- do.call(rbind, j1)
jGen$index <- row.names(jGen)

}
\seealso{
\url{https://www.w3.org/TR/prov-o/#wasGeneratedBy}
}

