% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInformedBy.R
\name{getInformedBy}
\alias{getInformedBy}
\title{Create a table of "wasInformedBy" relationships from PROV-JSON input.}
\usage{
getInformedBy(z)
}
\arguments{
\item{z}{A PROV-JSON data frame.}
}
\value{
A \code{data.frame} containing \code{inf}, \code{informed}, and \code{informant} fields.
}
\description{
\code{getInformedBy} returns a data frame containing the PROV-O "wasInformedBy" relationships extracted from a DDMoRe PROV-JSON document. Communication is the exchange of an entity by two activities, one activity using the entity generated by the other.
}
\examples{
### Reads in a sequence of JSON files from dir
loadOutput <- function(files) {
jList <- list()
for(f in files) {
  jList[[basename(f)]] <- fromJSON(txt = f)
}
jList
}

dir   <- "/data/jsonDocs"
jlist <- loadOutput(file.path(dir, dir(dir)))

j1 <- lapply(jlist, function(x) {
  z <- getInformedBy(x)
  z
})
jInf       <- do.call(rbind, j1)
jInf$index <- row.names(jInf)

}
\seealso{
\url{https://www.w3.org/TR/prov-o/#wasInformedBy}
}

