% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMentionOf.R
\name{getMentionOf}
\alias{getMentionOf}
\title{Create a table of "mentionOf" relationships from PROV-JSON input.}
\usage{
getMentionOf(z)
}
\arguments{
\item{z}{A PROV-JSON data frame.}
}
\value{
A \code{data.frame} containing \code{mentionOf}, \code{specificEntity}, \code{generalEntity}, and \code{bundle} fields.
}
\description{
\code{getMentionOf} returns a data frame containing the PROV-O "mentionOf" relationships extracted from a DDMoRe PROV-JSON document. A mention relates two entities with regard to a bundle. It is a special case of specialization. The mention of an entity in a bundle (containing a description of this entity) is another entity that is a specialization of the former and that presents at least the bundle as a further additional aspect.
}
\examples{
### Reads in a sequence of JSON files from dir
loadOutput <- function(files) {
jList <- list()
for(f in files) {
  jList[[basename(f)]] <- fromJSON(txt = f)
}
jList
}

dir   <- "/data/jsonDocs"
jlist <- loadOutput(file.path(dir, dir(dir)))

j1 <- lapply(jlist, function(x) {
  z <- getMentionOf(x)
  z
})
jMen       <- do.call(rbind, j1)
jMen$index <- row.names(jMen)

}
\seealso{
\url{https://www.w3.org/TR/prov-links/#mentionOf}
}

