% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUsed.R
\name{getUsed}
\alias{getUsed}
\title{Create a table of "used" relationships from PROV-JSON input.}
\usage{
getUsed(z)
}
\arguments{
\item{z}{A PROV-JSON data frame.}
}
\value{
A \code{data.frame} containing \code{used}, \code{activity}, and \code{entity} fields.
}
\description{
\code{getUsed} returns a data frame containing the PROV-O "used" relationships extracted from a DDMoRe PROV-JSON document. Usage is the beginning of utilizing an entity by an activity. Before usage, the activity had not begun to utilize this entity and could not have been affected by the entity.
}
\examples{
### Reads in a sequence of JSON files from dir
loadOutput <- function(files) {
jList <- list()
for(f in files) {
  jList[[basename(f)]] <- fromJSON(txt = f)
}
jList
}

dir   <- "/data/jsonDocs"
jlist <- loadOutput(file.path(dir, dir(dir)))

j1 <- lapply(jlist, function(x) {
  z <- getAssociatedWith(x)
  z
})
jUsed       <- do.call(rbind, j1)
jUsed$index <- row.names(jUsed)
}
\seealso{
\url{https://www.w3.org/TR/prov-o/#used}
}

